WEB/requirement.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.39.1
django-crispy-forms==1.7.2
gunicorn==19.9.0
idna==2.8
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.2
Pillow==6.0.0
psycopg2-binary==2.8.2
pylint==2.3.1
pyserial==3.4
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
urllib3==1.25.3
white==0.1.2
whitenoise==4.1.2
wrapt==1.11.1
